Add latest script versions from gitlab
